Move types and classes that can be reused across mobile and wear to a new common module
