+++ common/src/main/kotlin/com/boswelja/watchconnection/common/message/MessageReceiver.kt
@@ -1,4 +1,4 @@
-package com.boswelja.watchconnection.core.message
+package com.boswelja.watchconnection.common.message
 
 import android.content.BroadcastReceiver
 import android.content.Context
