+++ common/src/main/kotlin/com/boswelja/watchconnection/common/message/ReceivedMessage.kt
@@ -1,4 +1,4 @@
-package com.boswelja.watchconnection.core.message
+package com.boswelja.watchconnection.common.message
 
 import java.util.UUID
