+++ common/src/main/kotlin/com/boswelja/watchconnection/common/message/serialized/PrimitiveSerializers.kt
@@ -1,4 +1,4 @@
-package com.boswelja.watchconnection.core.message.serialized
+package com.boswelja.watchconnection.common.message.serialized
 
 import java.nio.ByteBuffer
