+++ common/src/test/kotlin/com/boswelja/watchconnection/common/message/ConcreteMessageReceiver.kt
@@ -1,4 +1,4 @@
-package com.boswelja.watchconnection.core.message
+package com.boswelja.watchconnection.common.message
 
 import android.content.Context
