+++ common/src/test/kotlin/com/boswelja/watchconnection/common/message/serialized/ConcreteDataType.kt
@@ -1,0 +1,5 @@
+package com.boswelja.watchconnection.common.message.serialized
+
+data class ConcreteDataType(
+    val data: String
+)
