+++ common/src/test/kotlin/com/boswelja/watchconnection/common/message/serialized/ConcreteMessageReceiver.kt
@@ -1,7 +1,7 @@
-package com.boswelja.watchconnection.core.message.serialized
+package com.boswelja.watchconnection.common.message.serialized
 
 import android.content.Context
-import com.boswelja.watchconnection.core.message.ReceivedMessage
+import com.boswelja.watchconnection.common.message.ReceivedMessage
 
 class ConcreteMessageReceiver : TypedMessageReceiver<ConcreteDataType>(
     serializer = ConcreteMessageSerializer
