+++ core/src/test/kotlin/com/boswelja/watchconnection/core/message/ConcreteMessagePlatform.kt
@@ -1,5 +1,7 @@
 package com.boswelja.watchconnection.core.message
 
+import com.boswelja.watchconnection.common.message.MessagePriority
+import com.boswelja.watchconnection.common.message.ReceivedMessage
 import java.util.UUID
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
