--- core/src/test/kotlin/com/boswelja/watchconnection/core/message/serialized/ConcreteDataType.kt
@@ -1,5 +1,0 @@
-package com.boswelja.watchconnection.core.message.serialized
-
-data class ConcreteDataType(
-    val data: String
-)
