+++ wearos/src/main/kotlin/com/boswelja/watchconnection/wearos/WearOSMessageReceiver.kt
@@ -1,7 +1,7 @@
 package com.boswelja.watchconnection.wearos
 
-import com.boswelja.watchconnection.core.Watch
-import com.boswelja.watchconnection.core.message.Messages
+import com.boswelja.watchconnection.common.Watch
+import com.boswelja.watchconnection.common.message.Messages
 import com.boswelja.watchconnection.wearos.Constants.WEAROS_PLATFORM
 import com.google.android.gms.wearable.MessageEvent
 import com.google.android.gms.wearable.WearableListenerService
